+++ changelog.md
@@ -2,7 +2,9 @@
 
 ### _2023.03.xx_ - 3.4.3
 - **Improved**: Accessibility: Header: Search: Dropdown.
+- **Improved**: Accessibility: Header: Mobile: Full Screen: Search.
 - **Updated**: Template: searchform.php
+- **Updated**: Template: partials/mobile/mobile-fullscreen-search.php
 
 ### _2023.02.22_ - 3.4.2
 - **Improved**: Customizer: Customizer panel scrollbar width for some browsers, like Chrome and Edge.